refactor api-based methods + tokenizer chat template
