+++ llmsanitize/model_contamination_checker.py
@@ -44,6 +44,7 @@ class ModelContaminationChecker(BaseContaminationChecker):
             openai_creds_key_file=self.openai_creds_key_file,
             local_port=self.local_port,
             local_api_type=self.local_api_type,
+            no_chat_template=self.no_chat_template,
             num_samples=self.num_samples,
             max_input_tokens=self.max_input_tokens,
             max_output_tokens=self.max_output_tokens,
@@ -82,6 +83,7 @@ class ModelContaminationChecker(BaseContaminationChecker):
             openai_creds_key_file=self.openai_creds_key_file,
             local_port=self.local_port,
             local_api_type=self.local_api_type,
+            no_chat_template=self.no_chat_template,
             num_samples=self.num_samples,
             max_input_tokens=self.max_input_tokens,
             max_output_tokens=self.max_output_tokens,
@@ -110,6 +112,7 @@ class ModelContaminationChecker(BaseContaminationChecker):
             openai_creds_key_file=self.openai_creds_key_file,
             local_port=self.local_port,
             local_api_type=self.local_api_type,
+            no_chat_template=self.no_chat_template,
             num_samples=self.num_samples,
             max_input_tokens=self.max_input_tokens,
             max_output_tokens=self.max_output_tokens,
