+++ llmsanitize/model_methods/cdd.py
@@ -66,6 +66,7 @@ def main_cdd(
     openai_creds_key_file: str = None,
     local_port: str = None,
     local_api_type: str = None,
+    no_chat_template: bool = False,
     num_samples: int = 20,
     max_input_tokens: int = 512,
     max_output_tokens: int = 128,
@@ -85,6 +86,7 @@ def main_cdd(
         openai_creds_key_file=openai_creds_key_file,
         local_port=local_port,
         local_api_type=local_api_type,
+        no_chat_template=no_chat_template,
         num_samples=1,
         max_input_tokens=max_input_tokens,
         max_output_tokens=max_output_tokens,
@@ -102,6 +104,7 @@ def main_cdd(
         openai_creds_key_file=openai_creds_key_file,
         local_port=local_port,
         local_api_type=local_api_type,
+        no_chat_template=no_chat_template,
         num_samples=num_samples,
         max_input_tokens=max_input_tokens,
         max_output_tokens=max_output_tokens,
