+++ llmsanitize/model_methods/guided_prompting.py
@@ -132,6 +132,7 @@ def main_guided_prompting(
     openai_creds_key_file: str = None,
     local_port: str = None,
     local_api_type: str = None,
+    no_chat_template: bool = False,
     num_samples: int = 1,
     max_input_tokens: int = 512,
     max_output_tokens: int = 128,
@@ -166,6 +167,7 @@ def main_guided_prompting(
         openai_creds_key_file=openai_creds_key_file,
         local_port=local_port,
         local_api_type=local_api_type,
+        no_chat_template=no_chat_template,
         num_samples=num_samples,
         max_input_tokens=max_input_tokens,
         max_output_tokens=max_output_tokens,
