+++ llmsanitize/model_methods/min_prob.py
@@ -152,6 +152,7 @@ def main_min_prob(
     openai_creds_key_file: str = None,
     local_port: str = None,
     local_api_type: str = None,
+    no_chat_template: bool = False,
     num_samples: int = 1,
     max_input_tokens: int = 512,
     max_output_tokens: int = 128,
@@ -176,6 +177,7 @@ def main_min_prob(
         openai_creds_key_file=openai_creds_key_file,
         local_port=local_port,
         local_api_type=local_api_type,
+        no_chat_template=no_chat_template,
         num_samples=num_samples,
         max_input_tokens=max_input_tokens,
         max_output_tokens=max_output_tokens,
@@ -193,6 +195,7 @@ def main_min_prob(
         openai_creds_key_file=openai_creds_key_file_2,
         local_port=local_port_2,
         local_api_type=local_api_type,
+        no_chat_template=no_chat_template,
         num_samples=num_samples,
         max_input_tokens=max_input_tokens,
         max_output_tokens=max_output_tokens,
