+++ scripts/tests/model/cdd/test_arc.sh
@@ -1,12 +1,10 @@
 # Get the options
-while getopts ":p:m:t:" option; do
+while getopts ":p:m:" option; do
    case $option in
       p) # port number
          port=$OPTARG;;
       m) # Enter model name
          model_name=$OPTARG;;
-      t) # local api type
-         local_api_type=$OPTARG;;
    esac
 done
 
@@ -16,7 +14,7 @@ echo "local port: ", $port
 # test guided prompting model contamination method
 python main.py \
 --eval_data_name allenai/ai2_arc \
---eval_data_config_name "ARC-Challenge" \
+--eval_data_config_name ARC-Challenge \
 --eval_set_key test \
 --text_key question \
 --label_key answerKey \
@@ -24,7 +22,6 @@ python main.py \
 --method cdd \
 --model_name $model_name \
 --local_port $port \
---local_api_type $local_api_type \
 --num_samples 20 \
 --temperature 0.8 \
 --cdd_alpha 0.05 \
