+++ scripts/tests/model/guided-prompting/openai/test_arc_openai.sh
