+++ scripts/tests/model/guided-prompting/openai/test_gsm8k_openai.sh
