+++ scripts/tests/model/guided-prompting/openai/test_hellaswag_openai.sh
