+++ scripts/tests/model/guided-prompting/openai/test_mmlu_openai.sh
