+++ scripts/tests/model/guided-prompting/openai/test_truthfulqa_openai.sh
