+++ scripts/tests/model/guided-prompting/openai/test_winogrande_openai.sh
