+++ scripts/tests/model/guided-prompting/test_hellaswag.sh
@@ -1,12 +1,10 @@
 # Get the options
-while getopts ":p:m:t:" option; do
+while getopts ":p:m:" option; do
    case $option in
       p) # port number
          port=$OPTARG;;
       m) # Enter model name
          model_name=$OPTARG;;
-      t) # local api type
-         local_api_type=$OPTARG;;
    esac
 done
 
@@ -17,13 +15,12 @@ echo "local port: ", $port
 python main.py \
 --eval_data_name Rowan/hellaswag \
 --eval_set_key validation \
---text_key "ctx" \
+--text_key ctx \
 --label_key activity_label \
 --n_eval_data_points 100 \
 --num_proc 40 \
 --method guided-prompting \
 --local_port $port \
---local_api_type $local_api_type \
 --model_name $model_name \
 --guided_prompting_task_type NLI \
 --use_local_model
