+++ scripts/tests/model/guided-prompting/test_truthfulqa.sh
@@ -1,12 +1,10 @@
 # Get the options
-while getopts ":p:m:t:" option; do
+while getopts ":p:m:" option; do
    case $option in
       p) # port number
          port=$OPTARG;;
       m) # Enter model name
          model_name=$OPTARG;;
-      t) # local api type
-         local_api_type=$OPTARG;;
    esac
 done
 
@@ -24,7 +22,6 @@ python main.py \
 --num_proc 40 \
 --method guided-prompting \
 --local_port $port \
---local_api_type $local_api_type \
 --model_name $model_name \
 --guided_prompting_task_type QA \
 --use_local_model
