+++ scripts/tests/model/guided-prompting/test_winogrande.sh
@@ -1,12 +1,10 @@
 # Get the options
-while getopts ":p:m:t:" option; do
+while getopts ":p:m:" option; do
    case $option in
       p) # port number
          port=$OPTARG;;
       m) # Enter model name
          model_name=$OPTARG;;
-      t) # local api type
-         local_api_type=$OPTARG;;
    esac
 done
 
@@ -23,7 +21,6 @@ python main.py \
 --n_eval_data_points 100 \
 --num_proc 40 \
 --method guided-prompting \
---local_api_type $local_api_type \
 --local_port $port \
 --model_name $model_name \
 --guided_prompting_task_type FIM \
