+++ scripts/tests/model/min-prob/test_gsm8k.sh
@@ -1,12 +1,10 @@
 # Get the options
-while getopts ":p:m:t:" option; do
+while getopts ":p:m:" option; do
    case $option in
       p) # port number
          port=$OPTARG;;
       m) # Enter model name
          model_name=$OPTARG;;
-      t) # local api type
-         local_api_type=$OPTARG;;
    esac
 done
 
@@ -16,9 +14,9 @@ echo "local port: ", $port
 # test min-K-prob model contamination method
 python main.py \
 --eval_data_name gsm8k \
---eval_data_config_name "main" \
+--eval_data_config_name main \
 --eval_set_key test \
---text_keys "question+answer" \
+--text_keys question+answer \
 --n_eval_data_points 100 \
 --num_proc 0 \
 --method min-prob \
