+++ scripts/tests/model/min-prob/test_mmlu.sh
@@ -1,12 +1,10 @@
 # Get the options
-while getopts ":p:m:t:" option; do
+while getopts ":p:m:" option; do
    case $option in
       p) # port number
          port=$OPTARG;;
       m) # Enter model name
          model_name=$OPTARG;;
-      t) # local api type
-         local_api_type=$OPTARG;;
    esac
 done
 
@@ -18,7 +16,7 @@ python main.py \
 --eval_data_name cais/mmlu \
 --eval_data_config_name all \
 --eval_set_key test \
---text_keys "question+choices+answer" \
+--text_keys question+choices+answer \
 --n_eval_data_points 100 \
 --num_proc 0 \
 --method min-prob \
