+++ scripts/tests/model/min-prob/test_truthfulqa.sh
@@ -1,12 +1,10 @@
 # Get the options
-while getopts ":p:m:t:" option; do
+while getopts ":p:m:" option; do
    case $option in
       p) # port number
          port=$OPTARG;;
       m) # Enter model name
          model_name=$OPTARG;;
-      t) # local api type
-         local_api_type=$OPTARG;;
    esac
 done
 
@@ -17,8 +15,8 @@ echo "local port: ", $port
 python main.py \
 --eval_data_name truthful_qa \
 --eval_data_config_name generation \
---eval_set_key "validation" \
---text_keys "question+best_answer" \
+--eval_set_key validation \
+--text_keys question+best_answer \
 --n_eval_data_points 100 \
 --num_proc 0 \
 --method min-prob \
