+++ scripts/tests/model/sharded-likelihood/test_truthfulqa.sh
@@ -13,7 +13,7 @@ python main.py \
 --eval_data_name truthful_qa \
 --eval_data_config_name generation \
 --eval_set_key validation \
---text_keys "question+best_answer" \
+--text_keys question+best_answer \
 --n_eval_data_points 100 \
 --num_proc 0 \
 --method sharded-likelihood \
