+++ scripts/tests/model/sharded-likelihood/test_winogrande.sh
@@ -13,7 +13,7 @@ python main.py \
 --eval_data_name winogrande \
 --eval_data_config_name winogrande_debiased \
 --eval_set_key test \
---text_keys "sentence+option1+option2+answer" \
+--text_keys sentence+option1+option2+answer \
 --n_eval_data_points 100 \
 --num_proc 0 \
 --method sharded-likelihood \
